+++ src/vs/workbench/contrib/terminal/common/terminal.ts
@@ -67,9 +67,12 @@ export interface ITerminalProfileService {
 	onDidChangeAvailableProfiles: Event<ITerminalProfile[]>;
 	getContributedDefaultProfile(shellLaunchConfig: IShellLaunchConfig): Promise<IExtensionTerminalProfile | undefined>;
 	registerContributedProfile(extensionIdentifier: string, id: string, title: string, options: ICreateContributedTerminalProfileOptions): Promise<void>;
-	//TODO: fix or move to a different location?
-	getContributedProfileProvider(extensionIdentifier: string, id: string): { createContributedTerminalProfile(options: ICreateContributedTerminalProfileOptions): Promise<void>; } | undefined;
-	registerTerminalProfileProvider(extensionIdentifier: string, id: string, profileProvider: { createContributedTerminalProfile(options: ICreateContributedTerminalProfileOptions): Promise<void>; }): IDisposable;
+	getContributedProfileProvider(extensionIdentifier: string, id: string): ITerminalProfileProvider | undefined;
+	registerTerminalProfileProvider(extensionIdentifier: string, id: string, profileProvider: ITerminalProfileProvider): IDisposable;
+}
+
+export interface ITerminalProfileProvider {
+	createContributedTerminalProfile(options: ICreateContributedTerminalProfileOptions): Promise<void>;
 }
 
 export interface IShellLaunchConfigResolveOptions {
