+++ src/vs/workbench/contrib/terminal/test/browser/terminalProfileService.test.ts
@@ -4,14 +4,14 @@
  *--------------------------------------------------------------------------------------------*/
 
 import { TestInstantiationService } from 'vs/platform/instantiation/test/common/instantiationServiceMock';
-import { ILocalTerminalService, ITerminalConfiguration } from 'vs/workbench/contrib/terminal/common/terminal';
+import { ILocalTerminalService, IOffProcessTerminalService, ITerminalConfiguration } from 'vs/workbench/contrib/terminal/common/terminal';
 import { TestConfigurationService } from 'vs/platform/configuration/test/common/testConfigurationService';
 import { TestExtensionService } from 'vs/workbench/test/common/workbenchTestServices';
 import { TerminalProfileService } from 'vs/workbench/contrib/terminal/browser/terminalProfileService';
 import { ITerminalContributionService } from 'vs/workbench/contrib/terminal/common/terminalExtensionPoints';
-import { IExtensionTerminalProfile, ITerminalProfile, ProfileSource } from 'vs/platform/terminal/common/terminal';
+import { IExtensionTerminalProfile, ITerminalProfile } from 'vs/platform/terminal/common/terminal';
 import { IRemoteTerminalService } from 'vs/workbench/contrib/terminal/browser/terminal';
-import { isWindows, OperatingSystem } from 'vs/base/common/platform';
+import { isLinux, isWindows, OperatingSystem } from 'vs/base/common/platform';
 import { MockContextKeyService } from 'vs/platform/keybinding/test/common/mockKeybindingService';
 import { IConfigurationService } from 'vs/platform/configuration/common/configuration';
 import { IExtensionService } from 'vs/workbench/services/extensions/common/extensions';
@@ -23,10 +23,15 @@ import { IRemoteAgentEnvironment } from 'vs/platform/remote/common/remoteAgentEn
 import { ThemeIcon } from 'vs/platform/theme/common/themeService';
 import { Codicon } from 'vs/base/common/codicons';
 import { equals } from 'vs/base/common/objects';
-import { strictEqual } from 'assert';
+import { deepStrictEqual, strictEqual } from 'assert';
 
 
 class TestTerminalProfileService extends TerminalProfileService {
+	hasRefreshedProfiles: Promise<void> | undefined;
+
+	override refreshAvailableProfiles(): void {
+		this.hasRefreshedProfiles = this._refreshAvailableProfilesNow();
+	}
 }
 
 
@@ -38,17 +43,7 @@ class TestTerminalContributionService implements ITerminalContributionService {
 	}
 }
 
-class TestRemoteTerminalService implements Partial<IRemoteTerminalService> {
-	private _profiles: ITerminalProfile[] = [];
-	async getProfiles(profiles: unknown, defaultProfile: unknown, includeDetectedProfiles?: boolean): Promise<ITerminalProfile[]> {
-		return this._profiles;
-	}
-	setProfiles(profiles: ITerminalProfile[]) {
-		this._profiles = profiles;
-	}
-}
-
-class TestLocalTerminalService implements Partial<ILocalTerminalService> {
+class TestOffProcessTerminalService implements Partial<IOffProcessTerminalService> {
 	private _profiles: ITerminalProfile[] = [];
 	async getProfiles(profiles: unknown, defaultProfile: unknown, includeDetectedProfiles?: boolean): Promise<ITerminalProfile[]> {
 		return this._profiles;
@@ -68,76 +63,7 @@ class TestRemoteAgentService implements Partial<IRemoteAgentService> {
 	}
 }
 
-const defaultTerminalConfig: Partial<ITerminalConfiguration> = {
-	profiles: {
-		windows: {
-			'PowerShell': {
-				'source': ProfileSource.Pwsh,
-				'icon': 'terminal-powershell'
-			},
-			'Command Prompt': {
-				'path': [
-					'${env:windir}\\Sysnative\\cmd.exe',
-					'${env:windir}\\System32\\cmd.exe'
-				],
-				'args': [],
-				'icon': 'terminal-cmd'
-			},
-			'Git Bash': {
-				'source': ProfileSource.GitBash
-			},
-		},
-		linux: {
-			'bash': {
-				'path': 'bash',
-				'icon': 'terminal-bash'
-			},
-			'zsh': {
-				'path': 'zsh'
-			},
-			'fish': {
-				'path': 'fish'
-			},
-			'tmux': {
-				'path': 'tmux',
-				'icon': 'terminal-tmux'
-			},
-			'pwsh': {
-				'path': 'pwsh',
-				'icon': 'terminal-powershell'
-			}
-		},
-		osx: {
-			'bash': {
-				'path': 'bash',
-				'args': [
-					'-l'
-				],
-				'icon': 'terminal-bash'
-			},
-			'zsh': {
-				'path': 'zsh',
-				'args': [
-					'-l'
-				]
-			},
-			'fish': {
-				'path': 'fish',
-				'args': [
-					'-l'
-				]
-			},
-			'tmux': {
-				'path': 'tmux',
-				'icon': 'terminal-tmux'
-			},
-			'pwsh': {
-				'path': 'pwsh',
-				'icon': 'terminal-powershell'
-			}
-		}
-	}
-};
+const defaultTerminalConfig: Partial<ITerminalConfiguration> = { profiles: { windows: {}, linux: {}, osx: {} } };
 
 suite('TerminalProfileService', () => {
 	let configurationService: TestConfigurationService;
@@ -151,8 +77,8 @@ suite('TerminalProfileService', () => {
 		let instantiationService = new TestInstantiationService();
 		let terminalContributionService = new TestTerminalContributionService();
 		let extensionService = new TestExtensionService();
-		let localTerminalService = new TestLocalTerminalService();
-		let remoteTerminalService = new TestRemoteTerminalService();
+		let localTerminalService = new TestOffProcessTerminalService();
+		let remoteTerminalService = new TestOffProcessTerminalService();
 		let contextKeyService = new MockContextKeyService();
 
 		instantiationService.stub(IContextKeyService, contextKeyService);
@@ -166,7 +92,7 @@ suite('TerminalProfileService', () => {
 		terminalProfileService = instantiationService.createInstance(TestTerminalProfileService);
 		localTerminalService.setProfiles([{
 			profileName: 'PowerShell',
-			path: 'C:Powershell.exe',
+			path: 'C:\\Powershell.exe',
 			isDefault: true,
 			icon: ThemeIcon.asThemeIcon(Codicon.terminalPowershell)
 		}]);
@@ -177,48 +103,51 @@ suite('TerminalProfileService', () => {
 			id: 'extension.js-debug.debugTerminal',
 			title: 'JavaScript Debug Terminal'
 		}]);
+		if (isWindows) {
+			remoteAgentService.setEnvironment(OperatingSystem.Windows);
+		} else if (isLinux) {
+			remoteAgentService.setEnvironment(OperatingSystem.Linux);
+		} else {
+			remoteAgentService.setEnvironment(OperatingSystem.Macintosh);
+		}
 	});
 
 	test('should filter out contributed profiles set to null', async () => {
-		if (isWindows) {
-			remoteAgentService.setEnvironment(OperatingSystem.Windows);
-			await configurationService.setUserConfiguration('terminal', {
-				integrated: {
-					profiles: {
-						windows: {
-							'JavaScript Debug Terminal': null
-						}
+		await configurationService.setUserConfiguration('terminal', {
+			integrated: {
+				profiles: {
+					windows: {
+						'JavaScript Debug Terminal': null
 					}
 				}
-			});
-			configurationService.onDidChangeConfigurationEmitter.fire({ affectsConfiguration: () => true } as any);
-			await terminalProfileService.refreshAvailableProfilesNow();
-			strictEqual(equals(terminalProfileService.availableProfiles, [{
-				profileName: 'PowerShell',
-				path: 'C:Powershell.exe',
-				isDefault: true,
-				icon: ThemeIcon.asThemeIcon(Codicon.terminalPowershell)
-			}]), true);
-			strictEqual(equals(terminalProfileService.contributedProfiles, []), true);
-		}
+			}
+		});
+		configurationService.onDidChangeConfigurationEmitter.fire({ affectsConfiguration: () => true } as any);
+		terminalProfileService.refreshAvailableProfiles();
+		await terminalProfileService.hasRefreshedProfiles;
+		deepStrictEqual(terminalProfileService.availableProfiles, [{
+			profileName: 'PowerShell',
+			path: 'C:\\Powershell.exe',
+			isDefault: true,
+			icon: ThemeIcon.asThemeIcon(Codicon.terminalPowershell)
+		}]);
+		strictEqual(equals(terminalProfileService.contributedProfiles, []), true);
 	});
 	test('should include contributed profiles', async () => {
 		configurationService.setUserConfiguration('terminal', { integrated: defaultTerminalConfig });
-		if (isWindows) {
-			remoteAgentService.setEnvironment(OperatingSystem.Windows);
-			await terminalProfileService.refreshAvailableProfilesNow();
-			strictEqual(equals(terminalProfileService.availableProfiles, [{
-				profileName: 'PowerShell',
-				path: 'C:Powershell.exe',
-				isDefault: true,
-				icon: ThemeIcon.asThemeIcon(Codicon.terminalPowershell)
-			}]), true);
-			strictEqual(equals(terminalProfileService.contributedProfiles, [{
-				extensionIdentifier: 'ms-vscode.js-debug-nightly',
-				icon: 'debug',
-				id: 'extension.js-debug.debugTerminal',
-				title: 'JavaScript Debug Terminal'
-			}]), true);
-		}
+		terminalProfileService.refreshAvailableProfiles();
+		await terminalProfileService.hasRefreshedProfiles;
+		strictEqual(equals(terminalProfileService.availableProfiles, [{
+			profileName: 'PowerShell',
+			path: 'C:\\Powershell.exe',
+			isDefault: true,
+			icon: ThemeIcon.asThemeIcon(Codicon.terminalPowershell)
+		}]), true);
+		strictEqual(equals(terminalProfileService.contributedProfiles, [{
+			extensionIdentifier: 'ms-vscode.js-debug-nightly',
+			icon: 'debug',
+			id: 'extension.js-debug.debugTerminal',
+			title: 'JavaScript Debug Terminal'
+		}]), true);
 	});
 });
